+++ src/components/select/select.component.spec.ts
@@ -21,8 +21,6 @@ import {OverlayHostService} from '../overlay-host/overlay-host.service';
 import {OverlayHost} from '../overlay-host/overlay-host.component';
 import {crossBrowserInitKeyboardEvent, KeyboardEventConfig} from '../../testing/keyboard-event';
 import {KeyCode} from '../../common/keycodes';
-import { ExpandOperator } from 'rxjs/operators/expand';
-
 
 describe('Select:', () => {
 
@@ -56,26 +54,27 @@ describe('Select:', () => {
 
     it('emits "clear" when the clear selection button is clicked',
             componentTest(() => TestComponent, `
-                <gtx-select clearable (clear)="onClear($event)"></gtx-select>`,
-                (fixture, testComponent) => {
+                <gtx-select clearable (change)="onChange($event)"></gtx-select>`,
+                (fixture, instance) => {
+                    instance.onChange = jasmine.createSpy('onChange');
                     fixture.detectChanges();
                     tick();
 
-                    expect(testComponent.onClear).not.toHaveBeenCalled();
+                    expect(instance.onChange).not.toHaveBeenCalled();
 
                     const clearButton = fixture.debugElement.query(By.css('gtx-button'));
                     clearButton.triggerEventHandler('click', document.createEvent('Event'));
                     tick();
                     fixture.detectChanges();
 
-                    expect(testComponent.onClear).toHaveBeenCalled();
+                    expect(instance.onChange).toHaveBeenCalledWith(null);
                 }
             )
     );
 
     it('updates the value via ngModel when the clear selection button is clicked',
             componentTest(() => TestComponent, `
-                <gtx-select [(ngModel)]="ngModelValue" (clear)="onClear($event)" clearable>
+                <gtx-select [(ngModel)]="ngModelValue" (change)="onChange($event)" clearable>
                         <gtx-option *ngFor="let option of options" [value]="option">{{ option }}</gtx-option>
                 </gtx-select>`,
                 (fixture, instance) => {
@@ -88,7 +87,7 @@ describe('Select:', () => {
                     tick();
                     fixture.detectChanges();
 
-                    expect(instance.ngModelValue).toEqual('');
+                    expect(instance.ngModelValue).toBeNull();
                 }
             )
         );
@@ -836,7 +835,6 @@ class TestComponent {
     testForm: FormGroup = new FormGroup({
         test: new FormControl('Bar')
     });
-    onClear = jasmine.createSpy('onClear');
 
     onBlur(): void {}
     onFocus(): void {}